feat(reg): settings column becomes staged accordion
