+++ apps/web/src/components/v2/registrations-panel.tsx
@@ -1,19 +1,21 @@
 "use client";
 
-// Organiser registration console (doc 16 §1.1, PROMPT-20a item 4):
-// settings (window, fee, capacity, bounded form-field builder), the
-// registration list with approve / waitlist / withdraw / refund, CSV export,
-// and the Stripe Connect onboarding banner that gates entry fees.
-import { useCallback, useEffect, useRef, useState } from "react";
+// Organiser registration console (doc 16 §1.1, PROMPT-20a item 4; PROMPT-52
+// reshape): composition shell owning state + API calls, rendering the
+// settings accordion (registration-settings.tsx) beside the registration
+// list. All derived numbers come from lib/registration-derive over the row
+// set this panel already loads.
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { apiV1, ApiV1Error } from "@/lib/client-v1";
 import { UpgradeGate } from "@/components/upgrade-gate";
-import { PlanBadge } from "@/components/plan-badge";
 import { useConfirm } from "@/components/ui/confirm-provider";
 import { Tip } from "@/components/ui/tip";
 import { msg } from "@/lib/messages";
 import { normalizeRefCode } from "@/lib/ref-code";
+import { registrationPulse } from "@/lib/registration-derive";
+import { RegistrationSettings } from "./registration-settings";
 
-interface FormField {
+export interface FormField {
   key: string;
   label: string;
   kind: "text" | "select" | "checkbox";
@@ -21,7 +23,7 @@ interface FormField {
   required: boolean;
 }
 
-interface Settings {
+export interface Settings {
   enabled: boolean;
   entrant_kind: "team" | "individual" | "pair";
   opens_at: string | null;
@@ -39,7 +41,7 @@ interface Settings {
   updated_at: string | null;
 }
 
-interface Registration {
+export interface Registration {
   id: string;
   status: string;
   ref_code: string | null;
@@ -53,6 +55,8 @@ interface Registration {
   payment_method: "offline" | "stripe" | null;
   payment_intent_id: string | null;
   refunded_cents: number;
+  /** Set once any refund ran — with refunded < amount it means incomplete. */
+  refunded_at: string | null;
   expires_at: string | null;
   offline_marked_paid_at: string | null;
   disputed_at: string | null;
@@ -99,21 +103,6 @@ function hoursLeft(iso: string | null): string {
   return h >= 48 ? "" : ` · ${h}h left`;
 }
 
-function toLocalInput(iso: string | null): string {
-  if (!iso) return "";
-  const d = new Date(iso);
-  const pad = (n: number) => String(n).padStart(2, "0");
-  return `${d.getFullYear()}-${pad(d.getMonth() + 1)}-${pad(d.getDate())}T${pad(d.getHours())}:${pad(d.getMinutes())}`;
-}
-
-function fromLocalInput(v: string): string | null {
-  return v ? new Date(v).toISOString() : null;
-}
-
-const CURRENCY_SYMBOLS: Record<string, string> = { gbp: "£", usd: "$", eur: "€" };
-function currencySymbol(code: string): string {
-  return CURRENCY_SYMBOLS[code?.toLowerCase()] ?? (code || "").toUpperCase();
-}
 
 export function RegistrationsPanel({
   divisionId,
@@ -153,6 +142,11 @@ export function RegistrationsPanel({
   // (pence). A local string keeps mid-typing states like "1." / "1.5" intact.
   const [feeText, setFeeText] = useState("");
   const feeInited = useRef(false);
+
+  // PROMPT-52 derived numbers — pure lenses over the already-loaded rows;
+  // no extra API round trip.
+  const capacity = settings?.capacity ?? null;
+  const pulse = useMemo(() => registrationPulse(regs, capacity), [regs, capacity]);
 
   const refresh = useCallback(async () => {
     const [s, r] = await Promise.all([
@@ -284,215 +278,21 @@ export function RegistrationsPanel({
   }
 
   const set = (patch: Partial<Settings>) => setSettings({ ...settings, ...patch });
-  const paidConfigured = settings.fee_cents > 0;
 
   return (
     <div className="grid gap-6 lg:grid-cols-[340px_1fr]">
       <aside className="space-y-4">
-        <section className="card space-y-3 p-5">
-          <h2 className="text-sm font-semibold text-slate-700">Registration settings</h2>
-
-          <label className="flex items-center gap-2 text-sm text-slate-700">
-            <input
-              type="checkbox"
-              disabled={!canEdit}
-              checked={settings.enabled}
-              onChange={(e) => set({ enabled: e.target.checked })}
-            />
-            Open for public registration
-          </label>
-
-          <label className="block text-xs text-slate-500">
-            Entrant type
-            <select
-              disabled={!canEdit}
-              value={settings.entrant_kind}
-              onChange={(e) => set({ entrant_kind: e.target.value as Settings["entrant_kind"] })}
-              className="input mt-1 w-full"
-            >
-              <option value="individual">Individual (creates a player)</option>
-              <option value="team">Team</option>
-              <option value="pair">Pair</option>
-            </select>
-          </label>
-
-          <div className="grid grid-cols-2 gap-2">
-            <label className="block text-xs text-slate-500">
-              Opens
-              <input
-                type="datetime-local"
-                disabled={!canEdit}
-                value={toLocalInput(settings.opens_at)}
-                onChange={(e) => set({ opens_at: fromLocalInput(e.target.value) })}
-                className="input mt-1 w-full"
-              />
-            </label>
-            <label className="block text-xs text-slate-500">
-              Closes
-              <input
-                type="datetime-local"
-                disabled={!canEdit}
-                value={toLocalInput(settings.closes_at)}
-                onChange={(e) => set({ closes_at: fromLocalInput(e.target.value) })}
-                className="input mt-1 w-full"
-              />
-            </label>
-          </div>
-
-          <label className="block text-xs text-slate-500">
-            Capacity (blank = uncapped; overflow joins the waitlist)
-            <input
-              type="number"
-              min={1}
-              disabled={!canEdit}
-              value={settings.capacity ?? ""}
-              onChange={(e) =>
-                set({ capacity: e.target.value === "" ? null : Number(e.target.value) })
-              }
-              className="input mt-1 w-full"
-            />
-          </label>
-
-          <div className="grid grid-cols-2 gap-2">
-            <label className="block text-xs text-slate-500">
-              <span className="flex items-center gap-1.5">
-                Entry fee ({currencySymbol(settings.currency)}; 0 = free)
-                {!paidAllowed && <PlanBadge feature="registration.paid" />}
-              </span>
-              <div className="mt-1 flex items-center">
-                <span className="rounded-l-md border border-r-0 border-slate-200 bg-slate-50 px-2.5 py-2 text-sm text-slate-500">
-                  {currencySymbol(settings.currency)}
-                </span>
-                <input
-                  type="text"
-                  inputMode="decimal"
-                  placeholder="0.00"
-                  disabled={!canEdit}
-                  value={feeText}
-                  onChange={(e) => {
-                    const t = e.target.value;
-                    // Allow empty, whole, or up to 2 decimals while typing.
-                    if (!/^\d*\.?\d{0,2}$/.test(t)) return;
-                    setFeeText(t);
-                    const pounds = parseFloat(t);
-                    set({ fee_cents: Number.isFinite(pounds) ? Math.round(pounds * 100) : 0 });
-                  }}
-                  onBlur={() => {
-                    // Normalise "1.5" → "1.50", "" → "0" on leaving the field.
-                    setFeeText((settings.fee_cents / 100).toFixed(2));
-                  }}
-                  className="input w-full rounded-l-none"
-                />
-              </div>
-            </label>
-            <label className="block text-xs text-slate-500">
-              Currency
-              <input
-                maxLength={3}
-                disabled={!canEdit}
-                value={settings.currency}
-                onChange={(e) => set({ currency: e.target.value.toLowerCase() })}
-                className="input mt-1 w-full"
-              />
-            </label>
-          </div>
-
-          {paidConfigured && (
-            <fieldset className="space-y-2">
-              <legend className="text-xs text-slate-500">How is the fee collected?</legend>
-              <label
-                className={`flex items-start gap-2.5 rounded-md border p-3 text-xs transition ${
-                  settings.payment_method === "offline"
-                    ? "border-purple-300 bg-purple-50"
-                    : "border-slate-200 bg-white hover:border-slate-300"
-                } ${canEdit ? "cursor-pointer" : ""}`}
-              >
-                <input
-                  type="radio"
-                  name="payment_method"
-                  className="mt-0.5"
-                  disabled={!canEdit}
-                  checked={settings.payment_method === "offline"}
-                  onChange={() => set({ payment_method: "offline" })}
-                />
-                <span>
-                  <span className="block font-medium text-slate-800">Pay the organiser</span>
-                  <span className="mt-0.5 block text-slate-500">
-                    Cash or bank transfer. Entries are pending until you mark them paid.
-                  </span>
-                </span>
-              </label>
-              {settings.payment_method === "offline" && (
-                <label className="block text-xs text-slate-500">
-                  Payment instructions for this division
-                  <textarea
-                    disabled={!canEdit}
-                    value={settings.payment_instructions ?? ""}
-                    onChange={(e) =>
-                      set({ payment_instructions: e.target.value || null })
-                    }
-                    rows={3}
-                    maxLength={2000}
-                    placeholder={
-                      settings.org_payment_instructions
-                        ? "Leave blank to use your organisation's instructions"
-                        : "e.g. bank details or “pay cash on the day” — or set organisation-wide instructions in Settings → Payments"
-                    }
-                    className="input mt-1 w-full font-mono text-xs"
-                  />
-                  {!settings.payment_instructions && settings.org_payment_instructions && (
-                    <span className="mt-1 block text-[11px] text-slate-400">
-                      Using your organisation&apos;s instructions.
-                    </span>
-                  )}
-                </label>
-              )}
-              <label
-                className={`flex items-start gap-2.5 rounded-md border p-3 text-xs transition ${
-                  settings.payment_method === "stripe"
-                    ? "border-purple-300 bg-purple-50"
-                    : "border-slate-200 bg-white hover:border-slate-300"
-                } ${!settings.charges_enabled ? "opacity-60" : canEdit ? "cursor-pointer" : ""}`}
-              >
-                <input
-                  type="radio"
-                  name="payment_method"
-                  className="mt-0.5"
-                  disabled={!canEdit || !settings.charges_enabled}
-                  checked={settings.payment_method === "stripe"}
-                  onChange={() => set({ payment_method: "stripe" })}
-                />
-                <span>
-                  <span className="flex items-center gap-1.5 font-medium text-slate-800">
-                    Card payment at sign-up
-                    {!paidAllowed && <PlanBadge feature="registration.paid" />}
-                  </span>
-                  <span className="mt-0.5 block text-slate-500">
-                    Paid via Stripe when they register · confirmed automatically · unpaid
-                    entries expire after 48h and the waitlist moves up.
-                  </span>
-                  {!settings.charges_enabled && (
-                    <a href="/settings/payments" className="mt-1 block font-medium text-purple-700 underline">
-                      Connect Stripe in Settings → Payments first
-                    </a>
-                  )}
-                </span>
-              </label>
-            </fieldset>
-          )}
-
-          {canEdit && (
-            <button type="button" disabled={busy} onClick={save} className="btn btn-primary w-full">
-              {busy ? "…" : "Save settings"}
-            </button>
-          )}
-          {saved && <p className="text-xs text-emerald-600">Saved.</p>}
-        </section>
-
-        <FormBuilder
-          fields={settings.form_fields}
+        <RegistrationSettings
+          settings={settings}
+          pulse={pulse}
+          feeText={feeText}
+          onFeeText={setFeeText}
+          onPatch={set}
+          onSave={save}
           canEdit={canEdit}
-          onChange={(form_fields) => set({ form_fields })}
+          paidAllowed={paidAllowed}
+          busy={busy}
+          saved={saved}
         />
 
         {paywall && <UpgradeGate feature={paywall} />}
@@ -635,129 +435,5 @@ export function RegistrationsPanel({
         )}
       </section>
     </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Bounded form-field builder (doc 16 §1.1): text / select / checkbox only.
-// ---------------------------------------------------------------------------
-
-function FormBuilder({
-  fields,
-  canEdit,
-  onChange,
-}: {
-  fields: FormField[];
-  canEdit: boolean;
-  onChange: (fields: FormField[]) => void;
-}) {
-  function update(i: number, patch: Partial<FormField>) {
-    onChange(fields.map((f, j) => (j === i ? { ...f, ...patch } : f)));
-  }
-
-  function add() {
-    // A blank label fails validation and blocks the whole save, so seed a
-    // valid default the organiser can rename.
-    const n = fields.length + 1;
-    onChange([...fields, { key: `question_${n}`, label: `Question ${n}`, kind: "text", required: false }]);
-  }
-
-  return (
-    <section className="card space-y-3 p-5">
-      <div className="flex items-center justify-between">
-        <div>
-          <h2 className="text-sm font-semibold text-slate-700">Extra sign-up questions</h2>
-          <p className="text-xs text-slate-400">
-            Ask registrants anything beyond name, email and date of birth.
-          </p>
-        </div>
-        {canEdit && fields.length < 12 && (
-          <button type="button" onClick={add} className="btn btn-ghost text-xs">
-            + Add question
-          </button>
-        )}
-      </div>
-      {fields.length === 0 && (
-        <p className="text-xs text-slate-400">
-          e.g. shirt size, dietary needs, emergency contact, club membership number, or a
-          &ldquo;I agree to the code of conduct&rdquo; checkbox.
-        </p>
-      )}
-      {fields.map((f, i) => (
-        <div key={i} className="space-y-2 rounded-md border border-slate-200 p-3">
-          <div className="flex gap-2">
-            <input
-              placeholder="Label"
-              disabled={!canEdit}
-              value={f.label}
-              onChange={(e) =>
-                update(i, {
-                  label: e.target.value,
-                  key:
-                    e.target.value
-                      .toLowerCase()
-                      .replace(/[^a-z0-9]+/g, "_")
-                      .replace(/^_+|_+$/g, "")
-                      .slice(0, 40) || f.key,
-                })
-              }
-              className="input flex-1 text-sm"
-            />
-            <select
-              disabled={!canEdit}
-              value={f.kind}
-              onChange={(e) =>
-                update(i, {
-                  kind: e.target.value as FormField["kind"],
-                  options: e.target.value === "select" ? (f.options ?? [""]) : undefined,
-                })
-              }
-              className="input text-sm"
-            >
-              <option value="text">Text</option>
-              <option value="select">Select</option>
-              <option value="checkbox">Checkbox</option>
-            </select>
-          </div>
-          {f.kind === "select" && (
-            <input
-              placeholder="Options, comma-separated"
-              disabled={!canEdit}
-              value={(f.options ?? []).join(", ")}
-              onChange={(e) =>
-                update(i, {
-                  options: e.target.value
-                    .split(",")
-                    .map((s) => s.trim())
-                    .filter(Boolean),
-                })
-              }
-              className="input w-full text-sm"
-            />
-          )}
-          <div className="flex items-center justify-between">
-            <label className="flex items-center gap-1.5 text-xs text-slate-500">
-              <input
-                type="checkbox"
-                disabled={!canEdit}
-                checked={f.required}
-                onChange={(e) => update(i, { required: e.target.checked })}
-              />
-              Required
-            </label>
-            {canEdit && (
-              <button
-                type="button"
-                onClick={() => onChange(fields.filter((_, j) => j !== i))}
-                className="text-xs text-red-500 hover:underline"
-              >
-                Remove
-              </button>
-            )}
-          </div>
-        </div>
-      ))}
-      <p className="text-[11px] text-slate-400">Save settings to apply form changes.</p>
-    </section>
   );
 }
